- Enabled PULL_DOWN switch instead of PULL_UP - extracted turn_on()/turn_off() functions - Moved starting delay calculations into setup() function because in debug mode they would do prints before serial Serial.begin() is called - Fixed the problem with buttons overshooting (by introducing a wait period before sending the stepper to sleep, this effectively acts as a brake for the button) - Fixed the problem of small amount of light when turned off (had to switch timers to 1khz phase corrected mode to do this, don't know why, maybe diodes were too slow for 4khz?) - Fixed the problem of starting with back light instead of front (position to level formula wasn't good) - Fixed the problem with long waits in choreographies (duration had to be switched to unsigned long)

+++ moonlight_types.h
@@ -36,9 +36,9 @@ const byte QUARTER = 96; // FULL_TURN / 4
 struct ControlPoint {
 	direction_t direction;
 	int position;
-	int duration;
+	unsigned long duration;
 
-	ControlPoint(direction_t dir, int pos, int dur) {
+	ControlPoint(direction_t dir, int pos, unsigned long dur) {
 		direction = dir;
 		duration = dur;
 		position = pos;
